+++ src/mono/System.Private.CoreLib/src/System/Object.Mono.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Runtime.CompilerServices;
-using System.Runtime.ExceptionServices;
 
 namespace System
 {
@@ -13,22 +12,6 @@ namespace System
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         protected extern object MemberwiseClone();
-    }
 
-    internal sealed class FinalizerHelper
-    {
-        [UnsafeAccessor(UnsafeAccessorKind.Method, Name = nameof(Finalize))]
-        private static extern void CallFinalize(object o);
-        private static void GuardedFinalize(object o)
-        {
-            try
-            {
-                CallFinalize(o);
-            }
-            catch (Exception ex) when (ExceptionHandling.IsHandledByGlobalHandler(ex))
-            {
-                // the handler returned "true" means the exception is now "handled" and we should continue.
-            }
-        }
     }
 }
